Move View into an extension.
